+++ TashaDesertsShop/TashaDesertsShop/App/Flows/Catalog/Cell/CakeTableViewCell.swift
@@ -16,14 +16,6 @@ class CakeTableViewCell: UITableViewCell {
     @IBOutlet weak var priceLabel: UILabel!
     @IBOutlet weak var productImage: RoundImageView!
     
-    override func awakeFromNib() {
-        super.awakeFromNib()
-    }
-
-    override func setSelected(_ selected: Bool, animated: Bool) {
-        super.setSelected(selected, animated: animated)
-    }
-    
     // MARK: Cell Configure
     
     func configure(with cellModel: CatalogResponseModel) {
@@ -34,27 +26,5 @@ class CakeTableViewCell: UITableViewCell {
         if let imageUrl = URL(string: cellModel.picUrl ?? "https://www.pngjoy.com/pngm/309/5828658_trailer-hd-omg-404-not-found-transparent-png.png"){
                     productImage.loadImage(url: imageUrl)
                 }
-    }
-}
-
-class RoundImageView: UIImageView {
-    
-    override func layoutSubviews() {
-        self.layer.cornerRadius = self.bounds.width / 2
-        let tap = UITapGestureRecognizer(target: self, action: #selector(tapFunction))
-        self.isUserInteractionEnabled = true
-        self.addGestureRecognizer(tap)
-    }
-    
-    @IBAction func tapFunction(sender: UITapGestureRecognizer) {
-        UIView.animate(withDuration: 0.75,
-                                    delay: 0,
-                                    usingSpringWithDamping: 0.25,
-                                    initialSpringVelocity: 0.75,
-                                    options: [.allowUserInteraction],
-                                    animations: {
-                                        self.bounds = self.bounds.insetBy(dx: 20, dy: 20)
-                                    },
-                                    completion: nil)
     }
 }
