+++ TashaDesertsShop/TashaDesertsShop/App/Flows/Catalog/Cell/CakeTableViewCell.xib
@@ -16,30 +16,30 @@
                 <rect key="frame" x="0.0" y="0.0" width="500" height="200"/>
                 <autoresizingMask key="autoresizingMask"/>
                 <subviews>
-                    <imageView clipsSubviews="YES" userInteractionEnabled="NO" contentMode="scaleAspectFit" horizontalHuggingPriority="251" verticalHuggingPriority="251" translatesAutoresizingMaskIntoConstraints="NO" id="fbk-1b-tJF">
-                        <rect key="frame" x="5" y="10" width="150" height="150"/>
+                    <imageView clipsSubviews="YES" userInteractionEnabled="NO" contentMode="scaleAspectFit" horizontalHuggingPriority="251" verticalHuggingPriority="251" translatesAutoresizingMaskIntoConstraints="NO" id="fbk-1b-tJF" customClass="RoundImageView" customModule="TashaDesertsShop" customModuleProvider="target">
+                        <rect key="frame" x="5" y="5" width="150" height="150"/>
                         <constraints>
                             <constraint firstAttribute="width" constant="150" id="f9c-tt-HHL"/>
                             <constraint firstAttribute="height" constant="150" id="qa0-PV-W2t"/>
                         </constraints>
                     </imageView>
                     <stackView opaque="NO" contentMode="scaleToFill" axis="vertical" translatesAutoresizingMaskIntoConstraints="NO" id="igW-xF-SYR">
-                        <rect key="frame" x="163" y="16" width="312" height="65"/>
+                        <rect key="frame" x="160" y="16" width="315" height="59"/>
                         <subviews>
                             <label opaque="NO" userInteractionEnabled="NO" contentMode="left" horizontalHuggingPriority="251" text="Name" textAlignment="natural" lineBreakMode="tailTruncation" numberOfLines="2" baselineAdjustment="alignBaselines" adjustsFontSizeToFit="NO" translatesAutoresizingMaskIntoConstraints="NO" id="60o-qa-h1P">
-                                <rect key="frame" x="0.0" y="0.0" width="312" height="26.5"/>
+                                <rect key="frame" x="0.0" y="0.0" width="315" height="26.5"/>
                                 <fontDescription key="fontDescription" type="boldSystem" pointSize="22"/>
                                 <nil key="textColor"/>
                                 <nil key="highlightedColor"/>
                             </label>
-                            <label opaque="NO" userInteractionEnabled="NO" contentMode="left" horizontalHuggingPriority="251" verticalHuggingPriority="251" text="Description" textAlignment="natural" lineBreakMode="tailTruncation" numberOfLines="2" baselineAdjustment="alignBaselines" adjustsFontSizeToFit="NO" translatesAutoresizingMaskIntoConstraints="NO" id="Agd-8b-sLB">
-                                <rect key="frame" x="0.0" y="26.5" width="312" height="18"/>
-                                <fontDescription key="fontDescription" type="system" pointSize="15"/>
+                            <label opaque="NO" userInteractionEnabled="NO" contentMode="left" horizontalHuggingPriority="251" verticalHuggingPriority="251" text="Description" textAlignment="natural" lineBreakMode="tailTruncation" numberOfLines="10" baselineAdjustment="alignBaselines" adjustsFontSizeToFit="NO" translatesAutoresizingMaskIntoConstraints="NO" id="Agd-8b-sLB">
+                                <rect key="frame" x="0.0" y="26.5" width="315" height="12"/>
+                                <fontDescription key="fontDescription" type="system" pointSize="10"/>
                                 <nil key="textColor"/>
                                 <nil key="highlightedColor"/>
                             </label>
                             <label opaque="NO" userInteractionEnabled="NO" contentMode="left" horizontalHuggingPriority="251" verticalHuggingPriority="251" text="price" textAlignment="natural" lineBreakMode="tailTruncation" baselineAdjustment="alignBaselines" adjustsFontSizeToFit="NO" translatesAutoresizingMaskIntoConstraints="NO" id="oDw-9t-18H">
-                                <rect key="frame" x="0.0" y="44.5" width="312" height="20.5"/>
+                                <rect key="frame" x="0.0" y="38.5" width="315" height="20.5"/>
                                 <fontDescription key="fontDescription" type="system" pointSize="17"/>
                                 <nil key="textColor"/>
                                 <nil key="highlightedColor"/>
@@ -48,9 +48,10 @@
                     </stackView>
                 </subviews>
                 <constraints>
+                    <constraint firstAttribute="bottom" secondItem="fbk-1b-tJF" secondAttribute="bottom" constant="5" id="7ZI-GB-czi"/>
                     <constraint firstItem="fbk-1b-tJF" firstAttribute="leading" secondItem="H2p-sc-9uM" secondAttribute="leading" constant="5" id="87n-5e-KE2"/>
-                    <constraint firstItem="igW-xF-SYR" firstAttribute="leading" secondItem="fbk-1b-tJF" secondAttribute="trailing" constant="8" symbolic="YES" id="AmM-cV-sg5"/>
-                    <constraint firstItem="fbk-1b-tJF" firstAttribute="top" secondItem="H2p-sc-9uM" secondAttribute="top" constant="10" id="Hyh-sM-VQA"/>
+                    <constraint firstItem="fbk-1b-tJF" firstAttribute="top" secondItem="H2p-sc-9uM" secondAttribute="top" constant="5" id="Hyh-sM-VQA"/>
+                    <constraint firstItem="igW-xF-SYR" firstAttribute="leading" secondItem="fbk-1b-tJF" secondAttribute="trailing" constant="5" id="Lbs-WX-bOW"/>
                     <constraint firstItem="igW-xF-SYR" firstAttribute="top" secondItem="H2p-sc-9uM" secondAttribute="topMargin" constant="5" id="P6R-kv-En8"/>
                     <constraint firstAttribute="trailingMargin" secondItem="igW-xF-SYR" secondAttribute="trailing" constant="5" id="QPt-6e-uzK"/>
                     <constraint firstAttribute="bottom" secondItem="fbk-1b-tJF" secondAttribute="bottom" constant="10" id="dne-Oc-pZB"/>
