+++ TashaDesertsShop/TashaDesertsShop/Core/UIcomponents/RoundImageView.swift
@@ -1,0 +1,32 @@
+//
+//  RoundImageView.swift
+//  TashaDesertsShop
+//
+//  Created by Евгений Старшов on 01.05.2022.
+//
+
+import Foundation
+import UIKit
+
+class RoundImageView: UIImageView {
+    
+    override func layoutSubviews() {
+        self.layer.cornerRadius = self.bounds.width / 2
+        let tap = UITapGestureRecognizer(target: self, action: #selector(tapFunction))
+        self.isUserInteractionEnabled = true
+        self.addGestureRecognizer(tap)
+    }
+    
+    @IBAction func tapFunction(sender: UITapGestureRecognizer) {
+        print("Tapped")
+        UIView.animate(withDuration: 0.75,
+                                    delay: 0,
+                                    usingSpringWithDamping: 0.25,
+                                    initialSpringVelocity: 0.75,
+                                    options: [.allowUserInteraction],
+                                    animations: {
+                                        self.bounds = self.bounds.insetBy(dx: 20, dy: 20)
+                                    },
+                                    completion: nil)
+    }
+}
